% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl.R
\name{dl_file}
\alias{dl_file}
\title{Download data from the JSA}
\usage{
dl_file(urls, file = tempfile(fileext = ".xlsx"))
}
\arguments{
\item{urls}{A vector of two (2!) URLs; the second is a fallback for the first}

\item{file}{Path to the file where the data should be saved}
}
\value{
Returns the full path to the file where the data is saved.
}
\description{
Download data from the JSA
}
