% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_reos.R, R/reos.R
\name{dl_reos}
\alias{dl_reos}
\title{Download data from the JSA Recruitment Experience & Outlook Survey}
\usage{
dl_reos(file = tempfile(fileext = ".xlsx"))

dl_reos(file = tempfile(fileext = ".xlsx"))
}
\arguments{
\item{file}{Path to the file where the data should be saved}
}
\value{
Returns the full path to the file where the data is saved.

Returns the full path to the file where the data is saved.
}
\description{
Download data from the JSA Recruitment Experience & Outlook Survey

Download data from the JSA Recruitment Experience & Outlook Survey
}
