% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reos.R
\name{possible_reos_urls}
\alias{possible_reos_urls}
\title{Given a date, identify two URLs for the JSA REOS data - corresponding to
the previous two months.}
\usage{
possible_reos_urls()
}
\value{
A vector of URLs, length 2
}
\description{
Given a date, identify two URLs for the JSA REOS data - corresponding to
the previous two months.
}
