% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivi.R
\name{read_individual_ivi_table}
\alias{read_individual_ivi_table}
\title{Read data from Jobs & Skills Australia's Internet Vacancy Index (IVI)}
\usage{
read_individual_ivi_table(file, table)
}
\arguments{
\item{file}{Path to the file where the data should be saved}

\item{tables}{A vector of tables to read, or "all" to read all tables.
Possible options are \verb{c("all", skill", "4dig", "2dig_states", "2dig_regions")}}
}
\value{
A tibble with the REOS data, in 'long'/tidy format
}
\description{
Read data from Jobs & Skills Australia's Internet Vacancy Index (IVI)
}
