% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivi.R
\name{read_ivi}
\alias{read_ivi}
\title{Read data from Jobs & Skills Australia's Internet Vacancy Index (IVI)}
\usage{
read_ivi(tables, path = tempdir())
}
\arguments{
\item{tables}{A vector of tables to read, or "all" to read all tables.
Possible options are \verb{c("all", skill", "4dig", "2dig_states", "2dig_regions")}}

\item{path}{Path to the directory where the file(s) should be saved}
}
\value{
A tibble with the REOS data, in 'long'/tidy format
}
\description{
Read data from Jobs & Skills Australia's Internet Vacancy Index (IVI)
}
\examples{

# Get a single table
read_ivi("4dig")

# Get multiple tables
read_ivi(c("4dig", "2dig_states"))

# Get all tables
read_ivi("all")

}
